+++ src/main/java/com/elfmcys/yesstevemodel/geckolib3/core/keyframe/bone/TransitionKeyFrame.java
@@ -6,13 +6,13 @@ import org.joml.Vector3f;
 public class TransitionKeyFrame extends BoneKeyFrame {
     private final Vector3v postPoint;
 
-    public TransitionKeyFrame(double firstStartTick, Vector3v firstPoint, Vector3v postPoint) {
+    public TransitionKeyFrame(float firstStartTick, Vector3v firstPoint, Vector3v postPoint) {
         super(0, firstStartTick, firstPoint);
         this.postPoint = postPoint;
     }
 
     @Override
-    public Vector3f getLerpPoint(ExpressionEvaluator<?> evaluator, double percentCompleted) {
+    public Vector3f getLerpPoint(ExpressionEvaluator<?> evaluator, float percentCompleted) {
         if (!isEnd(percentCompleted)) {
             return beginPoint.eval(evaluator);
         } else {
